separate story folder
